Fixed NPE when no lookup annotations used. Comments added.

+++ lookup.annotations/src/org/netbeans/modules/lookup/annotations/apt/NBServiceRegistrator.java
@@ -15,6 +15,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import javax.annotation.processing.AbstractProcessor;
@@ -29,6 +30,7 @@ import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeMirror;
+import javax.lang.model.util.SimpleAnnotationValueVisitor6;
 import javax.tools.Diagnostic.Kind;
 import javax.tools.FileObject;
 import javax.tools.JavaFileManager.Location;
@@ -49,6 +51,19 @@ import org.netbeans.modules.lookup.annotations.model.ServiceM;
 @SupportedAnnotationTypes(value = {"*"})
 @SupportedSourceVersion(SourceVersion.RELEASE_6)
 public class NBServiceRegistrator extends AbstractProcessor {
+    /**
+     * Annotation visitor for traversing all {@linkplain ContractProvided} annotations
+     * contained in a {@linkplain ContractsProvided} annotation
+     */
+    private class ContractsProvidedAnnotationVisitor extends SimpleAnnotationValueVisitor6<Void, Set<ContractM>> {
+        @Override
+        public Void visitAnnotation(AnnotationMirror a, Set<ContractM> contracts) {
+            processContractProvided(a, contracts);
+            return null;
+        }
+        
+    }
+    
     private Collection<ServiceM> globalServices = new ArrayList<ServiceM>();
     
     @Override
@@ -68,6 +83,26 @@ public class NBServiceRegistrator extends AbstractProcessor {
         loadMetaInfServices(persistedServices, StandardLocation.SOURCE_PATH);
         
         loadMetaInfServices(services, StandardLocation.CLASS_OUTPUT);
+        processAnnotatedServices(roundEnv, services);
+
+        invalidateMetaInfServices(services);
+        mergeServices(services, persistedServices);
+        
+        globalServices.addAll(services.values());
+        
+//        debugRegistrations(services);
+        
+        return true;
+    }
+
+    /**
+     * Goes through all classes annotated by {@linkplain Service} annotation
+     * and retrieves the contracts implemented
+     * 
+     * @param roundEnv {@link RoundEnvironment} instance
+     * @param services A bag to put the created services
+     */
+    private void processAnnotatedServices(RoundEnvironment roundEnv, Map<String, ServiceM> services) {
 
         for (Element element : roundEnv.getElementsAnnotatedWith(Service.class)) {
             if (element.getKind() == ElementKind.CLASS) {
@@ -83,17 +118,13 @@ public class NBServiceRegistrator extends AbstractProcessor {
                 service.getContracts().addAll(findInheritedContracts(serviceElement));
             }
         }
-
-        invalidateMetaInfServices(services);
-        mergeServices(services, persistedServices);
-        
-        globalServices.addAll(services.values());
-        
-//        debugRegistrations(services);
-        
-        return true;
     }
-
+    
+    /**
+     * A utility method to merge two bags of services
+     * @param toServices The bag to merge to
+     * @param fromServices The to merge from
+     */
     private void mergeServices(Map<String, ServiceM> toServices, Map<String, ServiceM> fromServices) {
         for(Map.Entry<String, ServiceM> entry : fromServices.entrySet()) {
             ServiceM currentService = toServices.get(entry.getKey());
@@ -105,6 +136,11 @@ public class NBServiceRegistrator extends AbstractProcessor {
         }
     }
     
+    /**
+     * This method checks the services from the bag against the actual state
+     * All services not comforming to the definition of a service are pruned from the bag
+     * @param services The bag of services
+     */
     private void invalidateMetaInfServices(Map<String, ServiceM> services) {
         for(Map.Entry<String, ServiceM> entry : services.entrySet()) {
             TypeElement serviceElement = processingEnv.getElementUtils().getTypeElement(entry.getKey());
@@ -116,16 +152,23 @@ public class NBServiceRegistrator extends AbstractProcessor {
         }        
     }
 
+    /**
+     * Loads services defined in a <i>META-INF/services</i> location
+     * @param services The services bag to load the new services to
+     * @param serviceLocation A {@link Location} specifying where to load the services from
+     */
     private void loadMetaInfServices(Map<String, ServiceM> services, Location serviceLocation) {
         try {
-            Field fManagerFld = processingEnv.getFiler().getClass().getDeclaredField("fileManager");
+            // ********** This is a hack to get hold of the source file location; there is no other way :( *******
+            Field fManagerFld = processingEnv.getFiler().getClass().getDeclaredField("fileManager"); // NOI18N
             fManagerFld.setAccessible(true);
 
             StandardJavaFileManager jfm = (StandardJavaFileManager) fManagerFld.get(processingEnv.getFiler());
+            // ********* End of hack *****************************************************************************
             for (File location : jfm.getLocation(serviceLocation)) {
                 if (location != null) {
                     processingEnv.getMessager().printMessage(Kind.NOTE, location.getAbsolutePath());
-                    loadMetaInfServices(services, location.getAbsolutePath() + File.separator + "META-INF" + File.separator + "services");
+                    loadMetaInfServices(services, location.getAbsolutePath() + File.separator + "META-INF" + File.separator + "services"); // NOI18N
                 }
             }
         } catch (Exception e) {
@@ -133,6 +176,12 @@ public class NBServiceRegistrator extends AbstractProcessor {
         }
     }
 
+    /**
+     * Loads services defined in a <i>META-INF/services</i> folder under the given path
+     * @param services The services bag to load the new services to
+     * @param srcPath The path to look for <i>META-INF/services</i>
+     * @throws java.io.IOException
+     */
     private void loadMetaInfServices(Map<String, ServiceM> services, String srcPath) throws IOException {
         File servicesFolder = new File(srcPath);
         if (!servicesFolder.exists() || !servicesFolder.isDirectory()) {
@@ -162,6 +211,11 @@ public class NBServiceRegistrator extends AbstractProcessor {
         }
     }
 
+    /**
+     * Finds all {@linkplain Contract}s deined up the class hierarchy starting by the given class
+     * @param serviceElement The class to start searching for inherited {@linkplain Contract}s
+     * @return Returns a set of inherited contract
+     */
     private Set<ContractM> findInheritedContracts(TypeElement serviceElement) {
         if (serviceElement == null) {
             return Collections.EMPTY_SET;
@@ -179,6 +233,11 @@ public class NBServiceRegistrator extends AbstractProcessor {
 
     }
 
+    /**
+     * Retrieves a set of contracts defined by {@linkplain ContractProvided} annotations on a class
+     * @param serviceElement The class to inspect
+     * @return Returns a set of defined contracts
+     */
     private Set<ContractM> findDefinedContracts(TypeElement serviceElement) {
         if (serviceElement == null) {
             return Collections.EMPTY_SET;
@@ -188,19 +247,20 @@ public class NBServiceRegistrator extends AbstractProcessor {
             TypeElement annElement = (TypeElement) annMirror.getAnnotationType().asElement();
             String annTypeName = processingEnv.getElementUtils().getBinaryName(annElement).toString();
 
-//            if (annTypeName.equals(ContractsProvided.class.getName())) {
-//                for (Map.Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : annMirror.getElementValues().entrySet()) {
-//                    if (entry.getKey().getSimpleName().toString().equals("value")) {
-//                        contracts.add(new ContractM(entry.getValue().toString().replace(".class", "")));
-//                    }
-//                }
-//            }
-            if (annTypeName.equals(ContractProvided.class.getName())) {
+            if (annTypeName.equals(ContractsProvided.class.getName())) {
                 for (Map.Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : annMirror.getElementValues().entrySet()) {
-                    if (entry.getKey().getSimpleName().toString().equals("value")) {
-                        contracts.add(new ContractM(entry.getValue().toString().replace(".class", "")));
+                    if (entry.getKey().getSimpleName().toString().equals("value")) { // NOI18N
+                        
+                        List<AnnotationValue> contractList = (List<AnnotationValue>)entry.getValue().getValue();
+                        for(AnnotationValue contr : contractList) {
+                            contr.accept(new ContractsProvidedAnnotationVisitor(), contracts);
+                        }
                     }
                 }
+            }
+            processingEnv.getMessager().printMessage(Kind.NOTE, annTypeName);
+            if (annTypeName.equals(ContractProvided.class.getName())) {
+                processContractProvided(annMirror, contracts);
             }
 
         }
@@ -209,44 +269,42 @@ public class NBServiceRegistrator extends AbstractProcessor {
 
     }
 
+    /**
+     * Debugging output
+     * @param services Bag of services
+     */
     private void debugRegistrations(Map<String, ServiceM> services) {
         for(Map.Entry<String, ServiceM> entry : services.entrySet()) {
-            processingEnv.getMessager().printMessage(Kind.NOTE, "Service: " + entry.getKey());
-            processingEnv.getMessager().printMessage(Kind.NOTE, "==============================================================");
+            processingEnv.getMessager().printMessage(Kind.NOTE, "Service: " + entry.getKey()); // NOI18N
+            processingEnv.getMessager().printMessage(Kind.NOTE, "=============================================================="); // NOI18N
             for(ContractM contract : entry.getValue().getContracts()) {
-                processingEnv.getMessager().printMessage(Kind.NOTE, "* " + contract.getContractClass());
+                processingEnv.getMessager().printMessage(Kind.NOTE, "* " + contract.getContractClass()); // NOI18N
             }
         }
     }
     
+    /**
+     * Writes the defined services to <i>META-INF/services</i> location
+     * @param services The services to write
+     * @throws java.io.IOException
+     */
     private void writeRegistrations(Collection<ServiceM> services) throws IOException {
-        Map<String, Set<String>> registrations = new HashMap<String, Set<String>>();
-        
-        for(ServiceM service : services) {
-            for(ContractM contract : service.getContracts()) {
-                Set<String> registeredServices = registrations.get(contract.getContractClass());
-                if (registeredServices == null) {
-                    registeredServices = new HashSet<String>();
-                    registrations.put(contract.getContractClass(), registeredServices);
-                }
-                registeredServices.add(service.getServiceClass());
-            }
-        }
+        Map<String, Set<String>> registrations = preprocessRegistrations(services);
         
         String servicesPath = null;
         for(Map.Entry<String, Set<String>> entry : registrations.entrySet()) {
-            final String METAINF = "META-INF/services";
+            final String METAINF = "META-INF/services"; // NOI18N
             String resourcePath = METAINF + "/" + entry.getKey();
             Writer writer = null;
             try {
-                FileObject fo = processingEnv.getFiler().createResource(StandardLocation.CLASS_OUTPUT, "", resourcePath, (Element) null);
+                FileObject fo = processingEnv.getFiler().createResource(StandardLocation.CLASS_OUTPUT, "", resourcePath, (Element) null); // NOI18N
                 if (servicesPath == null) {
                     servicesPath = fo.toUri().getPath();
                     servicesPath = servicesPath.substring(0, servicesPath.lastIndexOf(METAINF) + METAINF.length());
                 }
                 writer = fo.openWriter();
                 for(String serviceClass : entry.getValue()) {
-                    writer.write(serviceClass + "\n");
+                    writer.write(serviceClass + "\n"); // NOI18N
                 }
             } finally {
                 if (writer != null) {
@@ -256,12 +314,59 @@ public class NBServiceRegistrator extends AbstractProcessor {
                 }
             }
         }
-        
-        File servicesDir = new File(servicesPath);
-        if (servicesDir != null && servicesDir.isDirectory()) {
-            for(File regFile : servicesDir.listFiles()) {
-                if (!registrations.containsKey(regFile.getName())) {
-                    regFile.delete();
+        cleanupRegistrations(servicesPath, registrations);
+    }
+
+    /**
+     * A helper method to convert the service registrations from the internal
+     * representation to the one compatible with the <i>META-INF/services</i>
+     * registration paradigm.
+     * @param services The services to process
+     * @return Returns an in-memory representation of the <i>META-INF/services</i> registrations
+     */
+    private Map<String, Set<String>> preprocessRegistrations(Collection<ServiceM> services) {
+        Map<String, Set<String>> registrations = new HashMap<String, Set<String>>();
+        for (ServiceM service : services) {
+            for (ContractM contract : service.getContracts()) {
+                Set<String> registeredServices = registrations.get(contract.getContractClass());
+                if (registeredServices == null) {
+                    registeredServices = new HashSet<String>();
+                    registrations.put(contract.getContractClass(), registeredServices);
+                }
+                registeredServices.add(service.getServiceClass());
+            }
+        }
+        return registrations;
+    }
+    
+    /**
+     * A helper method to process the contents of {@linkplain ContractProvided} annotation
+     * @param annMirror The annotation mirror
+     * @param contracts The set of contracts to add the new contract to
+     */
+    private void processContractProvided(AnnotationMirror annMirror, Set<ContractM> contracts) {
+        for (Map.Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : annMirror.getElementValues().entrySet()) {
+            if (entry.getKey().getSimpleName().toString().equals("value")) {
+                contracts.add(new ContractM(entry.getValue().toString().replace(".class", "")));
+            }
+        }
+    }
+    
+    /**
+     * Checks the <i>META-INF/services</i> style registrations in the given path
+     * It removes all file-based registrations without their corresponding in-memory
+     * representation.
+     * @param servicesPath The path to look for the registrations
+     * @param registrations The in-memory registrations
+     */
+    private void cleanupRegistrations(String servicesPath, Map<String, Set<String>> registrations) {
+        if (servicesPath != null) {
+            File servicesDir = new File(servicesPath);
+            if (servicesDir != null && servicesDir.isDirectory()) {
+                for (File regFile : servicesDir.listFiles()) {
+                    if (!registrations.containsKey(regFile.getName())) {
+                        regFile.delete();
+                    }
                 }
             }
         }
